move all types into types.ts
