+++ convex/types.ts
@@ -1,12 +1,9 @@
 import { Infer, v } from 'convex/values';
+import { tableHelper } from './lib/utils';
 
-export const Message = v.object({
-  from: v.id('players'),
-  to: v.array(v.id('players')),
-  content: v.string(),
-  ts: v.number(),
-});
-export type Message = Infer<typeof Message>;
+// ts is milliseconds in game time
+const ts = v.number();
+export type GameTs = Infer<typeof ts>;
 
 // Hierarchical location within tree
 // TODO: build zone lookup from position, whether player-dependent or global.
@@ -20,42 +17,6 @@ export type Position = Infer<typeof Position>;
 
 export const Pose = v.object({ position: Position, orientation: v.number() });
 export type Pose = Infer<typeof Pose>;
-
-export const Stopped = v.object({
-  type: v.literal('stopped'),
-  reason: v.union(v.literal('interrupted'), v.literal('idle')),
-  pose: Pose,
-});
-
-export const Walking = v.object({
-  type: v.literal('walking'),
-  route: v.array(Position),
-  startTs: v.number(),
-  targetEndTs: v.number(),
-});
-
-export const Motion = v.union(Walking, Stopped);
-export type Motion = Infer<typeof Motion>;
-
-export const Player = v.object({
-  id: v.id('players'),
-  name: v.string(),
-  identity: v.string(),
-  motion: Motion,
-  thinking: v.boolean(),
-});
-export type Player = Infer<typeof Player>;
-
-export const Snapshot = v.object({
-  player: Player,
-  lastPlan: v.optional(v.object({ plan: v.string(), ts: v.number() })),
-  // recentMemories: v.array(memoryValidator),
-  nearbyPlayers: v.array(v.object({ player: Player, new: v.boolean() })),
-  nearbyConversations: v.array(
-    v.object({ conversationId: v.id('conversations'), messages: v.array(Message) }),
-  ),
-});
-export type Snapshot = Infer<typeof Snapshot>;
 
 export const Action = v.union(
   v.object({
@@ -81,3 +42,153 @@ export const Action = v.union(
   }),
 );
 export type Action = Infer<typeof Action>;
+
+export const Message = v.object({
+  from: v.id('players'),
+  to: v.array(v.id('players')),
+  content: v.string(),
+  ts,
+});
+export type Message = Infer<typeof Message>;
+
+export const Stopped = v.object({
+  type: v.literal('stopped'),
+  reason: v.union(v.literal('interrupted'), v.literal('idle')),
+  pose: Pose,
+});
+
+export const Walking = v.object({
+  type: v.literal('walking'),
+  route: v.array(Position),
+  startTs: v.number(),
+  targetEndTs: v.number(),
+});
+
+export const Motion = v.union(Walking, Stopped);
+export type Motion = Infer<typeof Motion>;
+
+export const Player = v.object({
+  id: v.id('players'),
+  name: v.string(),
+  identity: v.string(),
+  motion: Motion,
+  thinking: v.boolean(),
+  lastSpokeTs: v.number(),
+});
+export type Player = Infer<typeof Player>;
+
+export const Snapshot = v.object({
+  player: Player,
+  lastPlan: v.optional(v.object({ plan: v.string(), ts })),
+  // recentMemories: v.array(memoryValidator),
+  nearbyPlayers: v.array(
+    v.object({
+      player: Player,
+      new: v.boolean(),
+      // relationship: v.string()
+    }),
+  ),
+  nearbyConversations: v.array(
+    v.object({ conversationId: v.id('conversations'), messages: v.array(Message) }),
+  ),
+});
+export type Snapshot = Infer<typeof Snapshot>;
+
+// Journal documents are append-only, and define an player's state.
+export const Journal = tableHelper('journal', {
+  // TODO: maybe we can just use _creationTime?
+  ts,
+  playerId: v.id('players'),
+  // emojiSummary: v.string(),
+  data: v.union(
+    v.object({
+      type: v.literal('talking'),
+      // If they are speaking to a person in particular.
+      // If it's empty, it's just talking out loud.
+      audience: v.array(v.id('players')),
+      content: v.string(),
+      // Refers to the first message in the conversation.
+      conversationId: v.id('conversations'),
+    }),
+    Stopped,
+    Walking,
+    // When we run the agent loop.
+    v.object({
+      type: v.literal('planning'),
+      snapshot: Snapshot,
+    }),
+    // In case we don't do anything, confirm we're done planning.
+    v.object({
+      type: v.literal('continuing'),
+    }),
+
+    // Exercises left to the reader:
+
+    // v.object({
+    //   type: v.literal('thinking'),
+    // }),
+    // v.object({
+    //   type: v.literal('activity'),
+    //   description: v.string(),
+    // 	// objects: v.array(v.object({ id: v.id('objects'), action: v.string() })),
+    //   pose: Pose,
+    // }),
+  ),
+});
+export type Entry = Infer<typeof Journal.doc>;
+export type EntryType = Entry['data']['type'];
+export type EntryOfType<T extends EntryType> = Omit<Entry, 'data'> & {
+  data: Extract<Entry['data'], { type: T }>;
+};
+
+export const Memories = tableHelper('memories', {
+  playerId: v.id('players'),
+  description: v.string(),
+  embeddingId: v.id('embeddings'),
+  importance: v.number(),
+  ts,
+  data: v.union(
+    // Useful for seed memories, high level goals
+    v.object({
+      type: v.literal('identity'),
+    }),
+    // Setting up dynamics between players
+    v.object({
+      type: v.literal('relationship'),
+      playerId: v.id('players'),
+    }),
+    // Per-agent summary of recent observations
+    // Can start out all the same, but could be dependent on personality
+    v.object({
+      type: v.literal('conversation'),
+      conversationId: v.id('conversations'),
+    }),
+    v.object({
+      type: v.literal('plan'),
+    }),
+
+    // Exercises left to the reader:
+
+    // v.object({
+    //   type: v.literal('reflection'),
+    //   relatedMemoryIds: v.array(v.id('memories')),
+    // }),
+    // v.object({
+    //   type: v.literal('observation'),
+    //   object: v.string(),
+    //   pose: Pose,
+    // }),
+    // Seemed too noisey for every message for every party, but maybe?
+    // v.object({
+    //   type: v.literal('message'),
+    //   messageId: v.id('messages'),
+    //   relatedMemoryIds: v.optional(v.array(v.id('memories'))),
+    // }),
+    // Could be a way to have the agent reflect and change identities
+  ),
+});
+export type Memory = Infer<typeof Memories.doc>;
+export type MemoryType = Memory['data']['type'];
+export type MemoryOfType<T extends MemoryType> = Omit<Memory, 'data'> & {
+  data: Extract<Memory['data'], { type: T }>;
+};
